requirements.txt
# FastAPI & Server
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-multipart>=0.0.7

# Image Processing
pillow==10.4.0

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Configuration
pydantic-settings==2.1.0
